+++ mordant/src/jvmMain/kotlin/com/github/ajalt/mordant/internal/MppImpl.kt
@@ -103,7 +103,7 @@ internal actual fun readLineOrNullMpp(hideInput: Boolean): String? {
     if (hideInput) {
         val console = System.console()
         if (console != null) {
-            return console.readPassword().toString()
+            return console.readPassword().concatToString()
         }
     }
     return readlnOrNull()