+++ mordant/src/mingwX64Main/kotlin/com/github/ajalt/mordant/internal/tty.kt
@@ -1,6 +1,0 @@
-package com.github.ajalt.mordant.internal
-
-// hideInput is not currently implemented
-internal actual fun ttySetEcho(echo: Boolean) {
-
-}